# specs/003-fastapi-rag-integration/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: RAG Agent API
  description: API for interacting with the RAG (Retrieval-Augmented Generation) agent
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com

servers:
  - url: http://localhost:8000
    description: Local development server

paths:
  /query:
    post:
      summary: Query the RAG agent
      description: Send a query to the RAG agent and receive a response with supporting sources
      operationId: query_agent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QueryRequest'
      responses:
        '200':
          description: Successful response from the RAG agent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryResponse'
        '400':
          description: Bad request - invalid query format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /health:
    get:
      summary: Health check
      description: Check the health status of the RAG agent service
      operationId: health_check
      responses:
        '200':
          description: Health status information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthStatus'
        '500':
          description: Service unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    QueryRequest:
      type: object
      required:
        - query
      properties:
        query:
          type: string
          description: The user's question or query text
          example: "What is ROS 2 and how does it differ from ROS 1?"
          minLength: 1
          maxLength: 2000
        context:
          type: string
          description: Additional context to provide to the agent
          example: "I'm learning about robotics frameworks"
          maxLength: 5000
        user_id:
          type: string
          description: Identifier for the requesting user
          example: "user-123"
          pattern: "^[a-zA-Z0-9-_]+$"
        session_id:
          type: string
          description: Session identifier for conversation context
          example: "sess-456"
          pattern: "^[a-zA-Z0-9-_]+$"
        include_sources:
          type: boolean
          description: Whether to include source references in response
          default: true
          example: true

    QueryResponse:
      type: object
      required:
        - response
        - query_id
        - timestamp
        - status
      properties:
        response:
          type: string
          description: The agent's answer to the query
          example: "ROS 2 is the next-generation Robot Operating System..."
          minLength: 1
          maxLength: 10000
        sources:
          type: array
          description: References to documents used in the response
          items:
            $ref: '#/components/schemas/SourceReference'
          maxItems: 20
        query_id:
          type: string
          description: Unique identifier for the query
          example: "query-789"
          pattern: "^[a-zA-Z0-9-_]+$"
        timestamp:
          type: string
          format: date-time
          description: ISO 8601 timestamp of response
          example: "2025-12-26T10:00:00Z"
        execution_time:
          type: number
          description: Time taken to generate response in seconds
          example: 2.5
          minimum: 0
        status:
          type: string
          description: Status of the request
          enum: [success, error, partial]
          example: "success"

    SourceReference:
      type: object
      required:
        - id
        - title
        - relevance_score
      properties:
        id:
          type: string
          description: Unique identifier for the source
          example: "doc-001"
          minLength: 1
          maxLength: 100
        title:
          type: string
          description: Title of the source document
          example: "Introduction to ROS 2 Concepts"
          minLength: 1
          maxLength: 200
        url:
          type: string
          description: URL to the source
          format: uri
          example: "https://example.com/ros2-intro"
        relevance_score:
          type: number
          description: Relevance score (0-1)
          minimum: 0
          maximum: 1
          example: 0.85
        content_preview:
          type: string
          description: Preview of the relevant content
          maxLength: 500
          example: "ROS 2 is designed to address the limitations of ROS 1..."

    HealthStatus:
      type: object
      required:
        - status
        - timestamp
        - services
        - version
      properties:
        status:
          type: string
          description: Overall system status
          enum: [healthy, degraded, unavailable]
          example: "healthy"
        timestamp:
          type: string
          format: date-time
          description: ISO 8601 timestamp of status check
          example: "2025-12-26T10:00:00Z"
        services:
          type: object
          description: Status of individual services
          additionalProperties:
            type: object
            properties:
              status:
                type: string
                enum: [healthy, degraded, unavailable]
              details:
                type: string
                description: Additional service-specific details
          example:
            agent: { status: "healthy" }
            qdrant: { status: "healthy" }
            api: { status: "healthy" }
        version:
          type: string
          description: API version
          pattern: "^\\d+\\.\\d+\\.\\d+$"
          example: "1.0.0"

    ErrorResponse:
      type: object
      required:
        - error
        - code
        - timestamp
      properties:
        error:
          type: string
          description: Error message
          example: "Invalid query format"
          minLength: 1
          maxLength: 500
        code:
          type: string
          description: Error code
          pattern: "^[A-Z_]+$"
          example: "INVALID_QUERY"
        details:
          type: object
          description: Additional error details
          example:
            field: "query"
            reason: "Query is too short"
        timestamp:
          type: string
          format: date-time
          description: ISO 8601 timestamp of error
          example: "2025-12-26T10:00:00Z"